+++ backend/src/main/resources/db/migration/h2/V1.0__schema.sql
@@ -27,7 +27,7 @@ CREATE TABLE mooc_products
     id          BIGINT AUTO_INCREMENT NOT NULL,
     name        VARCHAR(255)          NOT NULL,
     description VARCHAR(255)          NOT NULL,
-    price       INT                   NOT NULL,
+    price       DECIMAL               NOT NULL,
     created_at  datetime              NULL,
     updated_at  datetime              NULL,
     CONSTRAINT pk_mooc_products PRIMARY KEY (id)
